pyproject.toml
[project]
name = "jj-rag-pipeline"
version = "0.1.0"
description = "RAG-powered recommendation system for Josh's laptop content"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sentence-transformers>=2.2.2",
    "langchain>=0.1.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.1",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "youtube-transcript-api>=0.6.0",
    "google-api-python-client>=2.100.0",
    "boto3>=1.42.43",
    "openai>=1.0.0",
    "google-genai>=0.3.0",
    "python-docx>=1.2.0",
]

[project.scripts]
jj = "cli:main"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
